% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pct_table.R
\name{get_pct_table}
\alias{get_pct_table}
\title{Get a summary table of percentages from a wide dataset}
\usage{
get_pct_table(.data)
}
\arguments{
\item{.data}{A wide-format data frame representing one row per census tract.
The data frame needs to have the columns: "number_reported", "rent_above30",
"rent_above50", and "tot_hh"}
}
\value{
A summary table of percentages (pct) and proportions (prop).
One row represents a household category (rent-burdened, severely burdened, and voucher participants)
}
\description{
\code{get_pct_table} gets a summary table of percentages of rent-burdened,
severely rent-burdened, and voucher participants from the total households
}
\examples{
acs_hud_de_geojoined \%>\%
 st_drop_geometry() \%>\%
 get_pct_table()

}
