% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_point.R
\name{main_point}
\alias{main_point}
\title{Draw a main point container with a main text, footnote, and an icon}
\usage{
main_point(main_text = list(), footnote = list(), icon = NULL)
}
\arguments{
\item{main_text}{A list of main text}

\item{footnote}{A list of footenote}

\item{icon}{An icon}
}
\value{
A UI definition in the shiny.tag class
}
\description{
Draw a main point container with a main text, footnote, and an icon
}
