% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_server}
\alias{map_server}
\title{Map Shiny Server module}
\usage{
map_server(id, selected_GEOIDs, geo_data)
}
\arguments{
\item{id}{}

\item{selected_GEOIDs}{}

\item{geo_data}{}
}
\value{

}
\description{
Map Shiny Server module
}
